@@ -4,15 +4,15 @@ apply plugin: 'com.github.dcendents.android-maven'
 group = "com.github.instagram"
 
 android {
-    compileSdkVersion 26
-    buildToolsVersion "26.0.1"
+    compileSdkVersion 29
+    buildToolsVersion "29.0.3"
 
     defaultConfig {
-        minSdkVersion 9
-        targetSdkVersion 26
+        minSdkVersion 15
+        targetSdkVersion 29
         versionCode 1
         versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
 
     }
     buildTypes {
@@ -32,13 +32,14 @@ task androidSourcesJar(type: Jar) {
 }
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile 'com.android.support:support-v4:26.+'
-    compile group: 'com.google.code.findbugs', name: 'jsr305', version: '3.0.0'
-    testCompile 'junit:junit:4.12'
-    testCompile "org.mockito:mockito-core:1.+"
-    testCompile "org.powermock:powermock-module-junit4:1.6.4"
-    testCompile "org.powermock:powermock-module-junit4-rule:1.6.4"
-    testCompile "org.powermock:powermock-api-mockito:1.6.4"
-    testCompile "org.powermock:powermock-classloading-xstream:1.6.4"
+    api fileTree(dir: 'libs', include: ['*.jar'])
+    //noinspection GradleCompatible
+    api 'androidx.legacy:legacy-support-v4:1.0.0'
+    api group: 'com.google.code.findbugs', name: 'jsr305', version: '3.0.2'
+    testImplementation 'junit:junit:4.13'
+    testImplementation "org.mockito:mockito-core:"
+    testImplementation "org.powermock:powermock-module-junit4:1.6.4"
+    testImplementation "org.powermock:powermock-module-junit4-rule:1.6.4"
+    testImplementation "org.powermock:powermock-api-mockito:1.6.4"
+    testImplementation "org.powermock:powermock-classloading-xstream:1.6.4"
 }
