@@ -11,10 +11,12 @@ package com.instagram.lazyload.base;
 
 import android.content.Context;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import androidx.fragment.app.Fragment;
+
 import javax.annotation.Nullable;
 
 /**
